requirements.txt
accelerate==1.3.0
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
filelock==3.17.0
Flask==3.1.0
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
Flask-SocketIO==5.5.1
fsspec==2024.12.0
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
IMAPClient==3.0.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
lxml==5.3.0
MarkupSafe==3.0.2
mpmath==1.3.0
mysql-connector-python==9.2.0
networkx==3.4.2
numpy==2.2.2
oauthlib==3.2.2
openai==1.60.1
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pluggy==1.5.0
proto-plus==1.25.0
protobuf==5.29.3
psutil==6.1.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
PyMuPDF==1.25.2
pyparsing==3.2.1
PyPDF2==3.0.1
PySocks==1.7.1
pytesseract==0.3.13
pytest==8.3.4
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2024.2
PyYAML==6.0.2
pyzmail36==1.0.5
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rpa==1.50.0
rsa==4.9
safetensors==0.5.2
setuptools==75.8.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sympy==1.13.1
tagui==1.50.0
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformers==4.48.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
wsproto==1.2.0